+++ src/se/fredsberg/euler/sequence/TriangleNumberSequence.java
@@ -1,17 +1,17 @@
 package se.fredsberg.euler.sequence;
 
 
-public class TriangleNumberSequence implements Sequence<Integer> {
+public class TriangleNumberSequence implements Sequence<Long> {
 
     /*
      * The n:th Triangle number is the sum of 1..n.
      */
     
-    int sum = 0;
+    long sum = 0;
     int i = 1;
 
     @Override
-    public Integer next() {
+    public Long next() {
         sum += i++;
         return sum;
     }
