+++ src/se/fredsberg/euler/series/SumOverFunctionSeries.java
@@ -3,22 +3,22 @@ package se.fredsberg.euler.series;
 import se.fredsberg.euler.function.Function;
 
 
-public class SumOverFunctionSeries implements Series<Integer> {
+public class SumOverFunctionSeries implements Series<Long> {
 
-    protected Integer sum = 0;
-    private final Function<Integer> function;
+    protected Long sum = 0L;
+    private final Function<Long> function;
     
-    public SumOverFunctionSeries(Function<Integer> function) {
+    public SumOverFunctionSeries(Function<Long> function) {
         this.function = function;
     }
 
     @Override
-    public void process(Integer i) {
+    public void process(Long i) {
         sum += function.calculate(i);
     }
 
     @Override
-    public Integer getResult() {
+    public Long getResult() {
         return sum;
     }
